backend/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.5
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.3
pylint==2.3.1
pymongo==3.7.2
python-dotenv==0.10.1
requests==2.17.1
six==1.12.0
typed-ast==1.3.1
urllib3==1.21.1
Werkzeug==0.15.1
wrapt==1.11.1
